refactor (encoding): move encoder to encoding package
